+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBStoreTimer.java
@@ -601,6 +601,8 @@ public class FDBStoreTimer extends StoreTimer {
         /** The number of query plans that include a {@link com.apple.foundationdb.record.query.plan.sorting.RecordQuerySortPlan}. */
         PLAN_SORT("number of sort plans", false),
         PLAN_DAM("number of dam plans", false),
+        /** The number of synthetic record type plans. */
+        PLAN_SYNTHETIC_TYPE("number of synthetic record types plans", false),
         /** The number of records given given to any filter within any plan. */
         QUERY_FILTER_GIVEN("number of records given to any filter within any plan", false),
         /** The number of records passed by any filter within any plan. */